move EnsembleModel from scratch to treemodel
